Add 5 structural metrics + fix HelmCT hue_sin2 parser (61 total)
New structural metrics (catches issues like Rec.2020 blue spike):

1. OOG Excursion — interpolate in-gamut pairs at 256 Lab steps,
   detect out-of-gamut intermediates (catches purple-shift artifacts)
2. Hue Reversal — reduce chroma from cusp to 0, detect hue direction
   changes (catches gamut mapping hue instability)
3. Near-Primary Hue Discontinuity — perturb sRGB/P3 primaries ±0.01,
   measure max hue angle jump (catches singularities)
4. Negative LMS — check 10K random sRGB colors for negative LMS
   values after M1 (catches improper cone response)
5. Extreme Chroma Stability — perturb P3/Rec.2020 primaries ±0.001
   in Lab, measure XYZ sensitivity (catches numerical instability)

Other fixes:
- HelmCT now reads both "hue_correction" list AND individual
  "hue_cos1/sin1/cos2/sin2" fields (was silently ignoring the latter)
- Gamut scan resolution increased to 300L×200C (was 150×120)
- Total: 61 metrics, 3038 gradient pairs, 3 gamuts

+++ core/gpu_metrics_perceptual.py
@@ -12,7 +12,7 @@ from .constants import (
 
 PI = math.pi
 
-_D65 = torch.tensor([0.95047, 1.0, 1.08883])
+_D65 = torch.tensor([0.95047, 1.0, 1.08883], dtype=torch.float64)
 _M_SRGB = torch.tensor([
     [0.4124564, 0.3575761, 0.1804375],
     [0.2126729, 0.7151522, 0.0721750],
